added valid pid check

+++ signals/8-signal_send.sh
@@ -7,8 +7,13 @@ if [ $# -ne 1 ]; then
 fi
 
 if [[ $1 =~ ^[0-9]+$ ]]; then
-	kill -3 $1
-	exit 0
+	kill -0 $1
+	if [ $? -e 0 ]; then
+		kill -3 $1
+		exit 0
+	else
+		exit 1
+	fi
 else
 	exit 1
 fi
